requirements.txt
boto3==1.24.71
requests==2.27.1
python-telegram-bot==13.11
